Reimplement ChainIndex::updateFromFilesystem to prevent segfaults
The problem was that original code could potentially read beyond memory-mapped file range and this
would segfault in rare cases when memory mapped range would be aligned with memory page boundary.

Specifically the problem was in blk00682.dat of bitcoin blockchain at this line:
https://github.com/citp/BlockSci/blob/393c6f4a31170c05210d6da025f8529657d976f5/src/parser/chain_index.cpp#L98

I took this opportunity and added some diagnostic error reporting as well.