fix: use larger address type if key exists
We previously relied on the loop being in ascending order to overwrite smaller values when there duplicate keys, which holds on MacOS but not on Ubuntu 16.04. On Ubuntu, when an address type wasn’t used (producing duplicate keys), the lower address type would be stored in the map, and as a result addresses were being serialized with the wrong address type. This would then lead to segfaults when trying to re-instantiate these addresses (e.g., when iterating over addresses in a cluster).

Fixes: #256, #273